+++ fairring/process_group.h
@@ -52,25 +52,25 @@ class ProcessGroupFairring : public c10d::ProcessGroup {
   class OptionsFairring : public c10d::ProcessGroup::Options {
    public:
     explicit OptionsFairring(
-        size_t maxMemoryAllocatedInBytes,
-        size_t maxPaddingAllocatedInBytes,
-        size_t minParallelism,
+        int64_t maxMemoryAllocatedInBytes,
+        int64_t maxPaddingAllocatedInBytes,
+        int64_t minParallelism,
         bool isHighPriorityStream = false,
         std::chrono::milliseconds timeout = kProcessGroupDefaultTimeout);
 
     ~OptionsFairring() override;
 
     static c10::intrusive_ptr<OptionsFairring> create(
-        size_t maxMemoryAllocatedInBytes,
-        size_t maxPaddingAllocatedInBytes,
-        size_t minParallelism,
+        int64_t maxMemoryAllocatedInBytes,
+        int64_t maxPaddingAllocatedInBytes,
+        int64_t minParallelism,
         bool isHighPriorityStream = false,
         std::chrono::milliseconds timeout = kProcessGroupDefaultTimeout);
 
    private:
-    const size_t maxMemoryAllocatedInBytes_;
-    const size_t maxPaddingAllocatedInBytes_;
-    const size_t minParallelism_;
+    const int64_t maxMemoryAllocatedInBytes_;
+    const int64_t maxPaddingAllocatedInBytes_;
+    const int64_t minParallelism_;
     const bool isHighPriorityStream_;
 
     friend ProcessGroupFairring;
@@ -176,9 +176,9 @@ class ProcessGroupFairring : public c10d::ProcessGroup {
  private:
   c10::intrusive_ptr<c10d::ProcessGroupNCCL> ncclPG_;
 
-  size_t maxMemoryAllocatedInBytes_;
-  size_t maxPaddingAllocatedInBytes_;
-  size_t minParallelism_;
+  int64_t maxMemoryAllocatedInBytes_;
+  int64_t maxPaddingAllocatedInBytes_;
+  int64_t minParallelism_;
   c10::intrusive_ptr<c10d::Store> store_;
   std::unique_ptr<MachineFairring> machine_;
 };
